+++ sdk/storage/azure-mgmt-storage/azure/mgmt/storage/v2019_04_01/models/_models.py
@@ -2412,6 +2412,13 @@ class StorageAccount(TrackedResource):
      disabled once it is enabled. Possible values include: "Disabled", "Enabled".
     :type large_file_shares_state: str or
      ~azure.mgmt.storage.v2019_04_01.models.LargeFileSharesState
+    :param allow_blob_public_access: Allow or disallow public access to all blobs or containers in
+     the storage account. The default interpretation is true for this property.
+    :type allow_blob_public_access: bool
+    :param minimum_tls_version: Set the minimum TLS version to be permitted on requests to storage.
+     The default interpretation is TLS 1.0 for this property. Possible values include: "TLS1_0",
+     "TLS1_1", "TLS1_2".
+    :type minimum_tls_version: str or ~azure.mgmt.storage.v2019_04_01.models.MinimumTlsVersion
     """
 
     _validation = {
@@ -2466,6 +2473,8 @@ class StorageAccount(TrackedResource):
         'geo_replication_stats': {'key': 'properties.geoReplicationStats', 'type': 'GeoReplicationStats'},
         'failover_in_progress': {'key': 'properties.failoverInProgress', 'type': 'bool'},
         'large_file_shares_state': {'key': 'properties.largeFileSharesState', 'type': 'str'},
+        'allow_blob_public_access': {'key': 'properties.allowBlobPublicAccess', 'type': 'bool'},
+        'minimum_tls_version': {'key': 'properties.minimumTlsVersion', 'type': 'str'},
     }
 
     def __init__(
@@ -2495,6 +2504,8 @@ class StorageAccount(TrackedResource):
         self.geo_replication_stats = None
         self.failover_in_progress = None
         self.large_file_shares_state = kwargs.get('large_file_shares_state', None)
+        self.allow_blob_public_access = kwargs.get('allow_blob_public_access', None)
+        self.minimum_tls_version = kwargs.get('minimum_tls_version', None)
 
 
 class StorageAccountCheckNameAvailabilityParameters(msrest.serialization.Model):
@@ -2578,6 +2589,13 @@ class StorageAccountCreateParameters(msrest.serialization.Model):
      disabled once it is enabled. Possible values include: "Disabled", "Enabled".
     :type large_file_shares_state: str or
      ~azure.mgmt.storage.v2019_04_01.models.LargeFileSharesState
+    :param allow_blob_public_access: Allow or disallow public access to all blobs or containers in
+     the storage account. The default interpretation is true for this property.
+    :type allow_blob_public_access: bool
+    :param minimum_tls_version: Set the minimum TLS version to be permitted on requests to storage.
+     The default interpretation is TLS 1.0 for this property. Possible values include: "TLS1_0",
+     "TLS1_1", "TLS1_2".
+    :type minimum_tls_version: str or ~azure.mgmt.storage.v2019_04_01.models.MinimumTlsVersion
     """
 
     _validation = {
@@ -2600,6 +2618,8 @@ class StorageAccountCreateParameters(msrest.serialization.Model):
         'enable_https_traffic_only': {'key': 'properties.supportsHttpsTrafficOnly', 'type': 'bool'},
         'is_hns_enabled': {'key': 'properties.isHnsEnabled', 'type': 'bool'},
         'large_file_shares_state': {'key': 'properties.largeFileSharesState', 'type': 'str'},
+        'allow_blob_public_access': {'key': 'properties.allowBlobPublicAccess', 'type': 'bool'},
+        'minimum_tls_version': {'key': 'properties.minimumTlsVersion', 'type': 'str'},
     }
 
     def __init__(
@@ -2620,6 +2640,8 @@ class StorageAccountCreateParameters(msrest.serialization.Model):
         self.enable_https_traffic_only = kwargs.get('enable_https_traffic_only', None)
         self.is_hns_enabled = kwargs.get('is_hns_enabled', None)
         self.large_file_shares_state = kwargs.get('large_file_shares_state', None)
+        self.allow_blob_public_access = kwargs.get('allow_blob_public_access', None)
+        self.minimum_tls_version = kwargs.get('minimum_tls_version', None)
 
 
 class StorageAccountKey(msrest.serialization.Model):
@@ -2781,6 +2803,13 @@ class StorageAccountUpdateParameters(msrest.serialization.Model):
      disabled once it is enabled. Possible values include: "Disabled", "Enabled".
     :type large_file_shares_state: str or
      ~azure.mgmt.storage.v2019_04_01.models.LargeFileSharesState
+    :param allow_blob_public_access: Allow or disallow public access to all blobs or containers in
+     the storage account. The default interpretation is true for this property.
+    :type allow_blob_public_access: bool
+    :param minimum_tls_version: Set the minimum TLS version to be permitted on requests to storage.
+     The default interpretation is TLS 1.0 for this property. Possible values include: "TLS1_0",
+     "TLS1_1", "TLS1_2".
+    :type minimum_tls_version: str or ~azure.mgmt.storage.v2019_04_01.models.MinimumTlsVersion
     """
 
     _attribute_map = {
@@ -2795,6 +2824,8 @@ class StorageAccountUpdateParameters(msrest.serialization.Model):
         'enable_https_traffic_only': {'key': 'properties.supportsHttpsTrafficOnly', 'type': 'bool'},
         'network_rule_set': {'key': 'properties.networkAcls', 'type': 'NetworkRuleSet'},
         'large_file_shares_state': {'key': 'properties.largeFileSharesState', 'type': 'str'},
+        'allow_blob_public_access': {'key': 'properties.allowBlobPublicAccess', 'type': 'bool'},
+        'minimum_tls_version': {'key': 'properties.minimumTlsVersion', 'type': 'str'},
     }
 
     def __init__(
@@ -2813,6 +2844,8 @@ class StorageAccountUpdateParameters(msrest.serialization.Model):
         self.enable_https_traffic_only = kwargs.get('enable_https_traffic_only', None)
         self.network_rule_set = kwargs.get('network_rule_set', None)
         self.large_file_shares_state = kwargs.get('large_file_shares_state', None)
+        self.allow_blob_public_access = kwargs.get('allow_blob_public_access', None)
+        self.minimum_tls_version = kwargs.get('minimum_tls_version', None)
 
 
 class StorageSkuListResult(msrest.serialization.Model):