+++ sdk/storage/azure-mgmt-storage/azure/mgmt/storage/v2019_06_01/models/_models_py3.py
@@ -3825,6 +3825,13 @@ class StorageAccount(TrackedResource):
     :type routing_preference: ~azure.mgmt.storage.v2019_06_01.models.RoutingPreference
     :ivar blob_restore_status: Blob restore status.
     :vartype blob_restore_status: ~azure.mgmt.storage.v2019_06_01.models.BlobRestoreStatus
+    :param allow_blob_public_access: Allow or disallow public access to all blobs or containers in
+     the storage account. The default interpretation is true for this property.
+    :type allow_blob_public_access: bool
+    :param minimum_tls_version: Set the minimum TLS version to be permitted on requests to storage.
+     The default interpretation is TLS 1.0 for this property. Possible values include: "TLS1_0",
+     "TLS1_1", "TLS1_2".
+    :type minimum_tls_version: str or ~azure.mgmt.storage.v2019_06_01.models.MinimumTlsVersion
     """
 
     _validation = {
@@ -3884,6 +3891,8 @@ class StorageAccount(TrackedResource):
         'private_endpoint_connections': {'key': 'properties.privateEndpointConnections', 'type': '[PrivateEndpointConnection]'},
         'routing_preference': {'key': 'properties.routingPreference', 'type': 'RoutingPreference'},
         'blob_restore_status': {'key': 'properties.blobRestoreStatus', 'type': 'BlobRestoreStatus'},
+        'allow_blob_public_access': {'key': 'properties.allowBlobPublicAccess', 'type': 'bool'},
+        'minimum_tls_version': {'key': 'properties.minimumTlsVersion', 'type': 'str'},
     }
 
     def __init__(
@@ -3897,6 +3906,8 @@ class StorageAccount(TrackedResource):
         is_hns_enabled: Optional[bool] = None,
         large_file_shares_state: Optional[Union[str, "LargeFileSharesState"]] = None,
         routing_preference: Optional["RoutingPreference"] = None,
+        allow_blob_public_access: Optional[bool] = None,
+        minimum_tls_version: Optional[Union[str, "MinimumTlsVersion"]] = None,
         **kwargs
     ):
         super(StorageAccount, self).__init__(tags=tags, location=location, **kwargs)
@@ -3925,6 +3936,8 @@ class StorageAccount(TrackedResource):
         self.private_endpoint_connections = None
         self.routing_preference = routing_preference
         self.blob_restore_status = None
+        self.allow_blob_public_access = allow_blob_public_access
+        self.minimum_tls_version = minimum_tls_version
 
 
 class StorageAccountCheckNameAvailabilityParameters(msrest.serialization.Model):
@@ -4013,6 +4026,13 @@ class StorageAccountCreateParameters(msrest.serialization.Model):
     :param routing_preference: Maintains information about the network routing choice opted by the
      user for data transfer.
     :type routing_preference: ~azure.mgmt.storage.v2019_06_01.models.RoutingPreference
+    :param allow_blob_public_access: Allow or disallow public access to all blobs or containers in
+     the storage account. The default interpretation is true for this property.
+    :type allow_blob_public_access: bool
+    :param minimum_tls_version: Set the minimum TLS version to be permitted on requests to storage.
+     The default interpretation is TLS 1.0 for this property. Possible values include: "TLS1_0",
+     "TLS1_1", "TLS1_2".
+    :type minimum_tls_version: str or ~azure.mgmt.storage.v2019_06_01.models.MinimumTlsVersion
     """
 
     _validation = {
@@ -4036,6 +4056,8 @@ class StorageAccountCreateParameters(msrest.serialization.Model):
         'is_hns_enabled': {'key': 'properties.isHnsEnabled', 'type': 'bool'},
         'large_file_shares_state': {'key': 'properties.largeFileSharesState', 'type': 'str'},
         'routing_preference': {'key': 'properties.routingPreference', 'type': 'RoutingPreference'},
+        'allow_blob_public_access': {'key': 'properties.allowBlobPublicAccess', 'type': 'bool'},
+        'minimum_tls_version': {'key': 'properties.minimumTlsVersion', 'type': 'str'},
     }
 
     def __init__(
@@ -4055,6 +4077,8 @@ class StorageAccountCreateParameters(msrest.serialization.Model):
         is_hns_enabled: Optional[bool] = None,
         large_file_shares_state: Optional[Union[str, "LargeFileSharesState"]] = None,
         routing_preference: Optional["RoutingPreference"] = None,
+        allow_blob_public_access: Optional[bool] = None,
+        minimum_tls_version: Optional[Union[str, "MinimumTlsVersion"]] = None,
         **kwargs
     ):
         super(StorageAccountCreateParameters, self).__init__(**kwargs)
@@ -4072,6 +4096,8 @@ class StorageAccountCreateParameters(msrest.serialization.Model):
         self.is_hns_enabled = is_hns_enabled
         self.large_file_shares_state = large_file_shares_state
         self.routing_preference = routing_preference
+        self.allow_blob_public_access = allow_blob_public_access
+        self.minimum_tls_version = minimum_tls_version
 
 
 class StorageAccountInternetEndpoints(msrest.serialization.Model):
@@ -4328,6 +4354,13 @@ class StorageAccountUpdateParameters(msrest.serialization.Model):
     :param routing_preference: Maintains information about the network routing choice opted by the
      user for data transfer.
     :type routing_preference: ~azure.mgmt.storage.v2019_06_01.models.RoutingPreference
+    :param allow_blob_public_access: Allow or disallow public access to all blobs or containers in
+     the storage account. The default interpretation is true for this property.
+    :type allow_blob_public_access: bool
+    :param minimum_tls_version: Set the minimum TLS version to be permitted on requests to storage.
+     The default interpretation is TLS 1.0 for this property. Possible values include: "TLS1_0",
+     "TLS1_1", "TLS1_2".
+    :type minimum_tls_version: str or ~azure.mgmt.storage.v2019_06_01.models.MinimumTlsVersion
     """
 
     _attribute_map = {
@@ -4343,6 +4376,8 @@ class StorageAccountUpdateParameters(msrest.serialization.Model):
         'network_rule_set': {'key': 'properties.networkAcls', 'type': 'NetworkRuleSet'},
         'large_file_shares_state': {'key': 'properties.largeFileSharesState', 'type': 'str'},
         'routing_preference': {'key': 'properties.routingPreference', 'type': 'RoutingPreference'},
+        'allow_blob_public_access': {'key': 'properties.allowBlobPublicAccess', 'type': 'bool'},
+        'minimum_tls_version': {'key': 'properties.minimumTlsVersion', 'type': 'str'},
     }
 
     def __init__(
@@ -4360,6 +4395,8 @@ class StorageAccountUpdateParameters(msrest.serialization.Model):
         network_rule_set: Optional["NetworkRuleSet"] = None,
         large_file_shares_state: Optional[Union[str, "LargeFileSharesState"]] = None,
         routing_preference: Optional["RoutingPreference"] = None,
+        allow_blob_public_access: Optional[bool] = None,
+        minimum_tls_version: Optional[Union[str, "MinimumTlsVersion"]] = None,
         **kwargs
     ):
         super(StorageAccountUpdateParameters, self).__init__(**kwargs)
@@ -4375,6 +4412,8 @@ class StorageAccountUpdateParameters(msrest.serialization.Model):
         self.network_rule_set = network_rule_set
         self.large_file_shares_state = large_file_shares_state
         self.routing_preference = routing_preference
+        self.allow_blob_public_access = allow_blob_public_access
+        self.minimum_tls_version = minimum_tls_version
 
 
 class StorageQueue(Resource):